added geometry preprocessing and moved LBI out of helpers
