+++ src/util/globals.js
@@ -48,6 +48,7 @@ export const legendSpacing = 4;
 export const nonTipNodeRadius = 0;
 export const tipRadius = 4;
 export const tipRadiusOnLegendMatchMultiplier = 1.7;
+export const defaultDistanceMeasures = ['num_date', 'div'];
 export const HIColorDomain = genericDomain.map((d) => {
   return Math.round(100 * (d * 3.6)) / 100;
 });
